requirements.txt
html-testRunner
coverage
faker
pandas
indic_transliteration
icecream
sklearn
numpy
datasketch
openai
langchain
langchain-openai
beautifulsoup4
faiss-cpu
python-dotenv
macnotesapp
unstructured
pypdf 
tkhtmlview
yake
pypdf2
pdfplumber
anytree
langchain_community
colorama
dedupe
unidecode